+++ docs/api.rst
@@ -3,9 +3,14 @@ API
 
 .. autosummary::
    UFS2ARCO.FV3Dataset
+   UFS2ARCO.MOM6Dataset
 
 .. autoclass:: UFS2ARCO.ufsdataset.UFSDataset
 
 .. autoclass:: UFS2ARCO.FV3Dataset
+   :members:
+   :inherited-members:
+
+.. autoclass:: UFS2ARCO.MOM6Dataset
    :members:
    :inherited-members:
